requirements.txt
flask
requests
beautifulsoup4
line-bot-sdk
schedule


